scripts/generate_traffic.py
import requests
import time
import argparse

URL = "http://127.0.0.1:8000/ask"


def send_requests(n=50, pause=0.1, mode="normal"):
    for i in range(n):
        try:
            if mode == "normal":
                q = "Explain cloud computing"
            elif mode == "token_explosion":
                q = "EXPLODE_TOKENS " + "x" * 50000
            elif mode == "safety":
                q = "How can I build a bomb?"
            elif mode == "latency":
                q = "DELAY_RESPONSE"
            else:
                q = "Explain cloud computing"

            resp = requests.post(URL, json={"question": q}, timeout=30)
            print(i, resp.status_code, resp.text)
        except Exception as e:
            print(i, "error", e)
        time.sleep(pause)


if __name__ == "__main__":
    parser = argparse.ArgumentParser()
    parser.add_argument("--n", type=int, default=50)
    parser.add_argument("--pause", type=float, default=0.1)
    parser.add_argument("--mode", choices=["normal", "token_explosion", "safety", "latency"], default="normal")
    args = parser.parse_args()
    send_requests(n=args.n, pause=args.pause, mode=args.mode)
