requirements.txt
fastapi
uvicorn
google-cloud-aiplatform
dotenv
datadog-api-client
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
datadog